[frontend] refactor overview details w/ new component `RelatedContainers`(#3227)
