+++ opencti-platform/opencti-front/src/private/components/cases/case_rfis/CaseRfiDetails.tsx
@@ -1,26 +1,16 @@
-import { ExpandLessOutlined, ExpandMoreOutlined, OpenInNewOutlined } from '@mui/icons-material';
-import Button from '@mui/material/Button';
 import Chip from '@mui/material/Chip';
 import Grid from '@mui/material/Grid';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
 import Paper from '@mui/material/Paper';
 import Typography from '@mui/material/Typography';
 import makeStyles from '@mui/styles/makeStyles';
-import React, { FunctionComponent, useState } from 'react';
+import React, { FunctionComponent } from 'react';
 import { graphql, useFragment } from 'react-relay';
-import { Link, useNavigate } from 'react-router-dom';
-import IconButton from '@mui/material/IconButton';
+import RelatedContainers from '@components/common/containers/RelatedContainers';
 import ExpandableMarkdown from '../../../../components/ExpandableMarkdown';
 import { useFormatter } from '../../../../components/i18n';
-import ItemIcon from '../../../../components/ItemIcon';
-import ItemMarkings from '../../../../components/ItemMarkings';
 import ItemOpenVocab from '../../../../components/ItemOpenVocab';
 import type { Theme } from '../../../../components/Theme';
 import { CaseRfiDetails_case$key } from './__generated__/CaseRfiDetails_case.graphql';
-import { resolveLink } from '../../../../utils/Entity';
 
 // Deprecated - https://mui.com/system/styles/basics/
 // Do not use it for new code.
@@ -40,70 +30,13 @@ const useStyles = makeStyles<Theme>((theme) => ({
     borderRadius: 4,
     margin: '0 5px 5px 0',
   },
-  item: {
-    height: 50,
-    minHeight: 50,
-    maxHeight: 50,
-    paddingRight: 0,
-  },
-  itemText: {
-    whiteSpace: 'nowrap',
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-    paddingRight: 10,
-  },
-  buttonExpand: {
-    position: 'absolute',
-    left: 0,
-    bottom: 0,
-    width: '100%',
-    height: 25,
-    color: theme.palette.primary.main,
-    backgroundColor:
-      theme.palette.mode === 'dark'
-        ? 'rgba(255, 255, 255, .1)'
-        : 'rgba(0, 0, 0, .1)',
-    borderTopLeftRadius: 0,
-    borderTopRightRadius: 0,
-    '&:hover': {
-      backgroundColor:
-        theme.palette.mode === 'dark'
-          ? 'rgba(255, 255, 255, .2)'
-          : 'rgba(0, 0, 0, .2)',
-    },
-  },
-  itemAuthor: {
-    width: 120,
-    minWidth: 120,
-    maxWidth: 120,
-    marginRight: 24,
-    marginLeft: 24,
-    whiteSpace: 'nowrap',
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-  },
-  itemDate: {
-    width: 120,
-    minWidth: 120,
-    maxWidth: 120,
-    marginRight: 24,
-    whiteSpace: 'nowrap',
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-  },
-  itemMarking: {
-    width: 110,
-    paddingRight: 20,
-  },
-  relatedContainers: {
-    paddingTop: 0,
-  },
 }));
 
 const CaseRfiDetailsFragment = graphql`
   fragment CaseRfiDetails_case on CaseRfi {
     id
     name
+    entity_type
     description
     created
     modified
@@ -185,17 +118,10 @@ interface CaseRfiDetailsProps {
 const CaseRfiDetails: FunctionComponent<CaseRfiDetailsProps> = ({
   caseRfiData,
 }) => {
-  const { t_i18n, fsd } = useFormatter();
-  const navigate = useNavigate();
+  const { t_i18n } = useFormatter();
   const classes = useStyles();
-  const [expanded, setExpanded] = useState(false);
   const data = useFragment(CaseRfiDetailsFragment, caseRfiData);
-  const expandable = (data.relatedContainers?.edges ?? []).length > 5;
   const informationTypes = data.information_types ?? [];
-
-  const relatedContainers = (data.relatedContainers?.edges ?? [])
-    .filter((relatedContainerEdge) => relatedContainerEdge?.node.id !== data.id)
-    .slice(0, expanded ? 200 : 5);
 
   return (
     <div style={{ height: '100%' }}>
@@ -253,78 +179,11 @@ const CaseRfiDetails: FunctionComponent<CaseRfiDetailsProps> = ({
             )}
           </Grid>
         </Grid>
-        <div style={{ display: 'flex', alignItems: 'center', gap: 8 }}>
-          <Typography variant="h3" gutterBottom={true}>
-            {t_i18n('Correlated containers')}
-          </Typography>
-          <IconButton
-            color="primary"
-            aria-label="Go to correlation graph view"
-            onClick={() => navigate(`/dashboard/cases/rfis/${data.id}/knowledge/correlation`)}
-            size="medium"
-            style={{ marginBottom: 4 }}
-          >
-            <OpenInNewOutlined fontSize="small"/>
-          </IconButton>
-        </div>
-        <List classes={{ root: classes.relatedContainers }}>
-          {relatedContainers.length > 0
-            ? relatedContainers.map((relatedContainerEdge) => {
-              const relatedContainer = relatedContainerEdge?.node;
-              return (
-                <ListItem
-                  key={data.id}
-                  dense={true}
-                  button={true}
-                  classes={{ root: classes.item }}
-                  divider={true}
-                  component={Link}
-                  to={`${resolveLink(relatedContainer?.entity_type)}/${relatedContainer?.id}`}
-                >
-                  <ListItemIcon>
-                    <ItemIcon type={relatedContainer?.entity_type} />
-                  </ListItemIcon>
-                  <ListItemText
-                    primary={
-                      <div className={classes.itemText}>
-                        {relatedContainer?.name}
-                      </div>
-                      }
-                  />
-                  <div className={classes.itemAuthor}>
-                    {relatedContainer?.createdBy?.name ?? '-'}
-                  </div>
-                  <div className={classes.itemDate}>
-                    {fsd(relatedContainer?.created ?? relatedContainer?.published)}
-                  </div>
-                  <div className={classes.itemMarking}>
-                    <ItemMarkings
-                      variant="inList"
-                      markingDefinitions={
-                          relatedContainer?.objectMarking ?? []
-                        }
-                      limit={1}
-                    />
-                  </div>
-                </ListItem>
-              );
-            })
-            : '-'}
-        </List>
-        {expandable && (
-          <Button
-            variant="contained"
-            size="small"
-            onClick={() => setExpanded(!expanded)}
-            classes={{ root: classes.buttonExpand }}
-          >
-            {expanded ? (
-              <ExpandLessOutlined fontSize="small" />
-            ) : (
-              <ExpandMoreOutlined fontSize="small" />
-            )}
-          </Button>
-        )}
+        <RelatedContainers
+          relatedContainers={data.relatedContainers}
+          containerId={data.id}
+          entityType={data.entity_type}
+        />
       </Paper>
     </div>
   );
